Cost testing for first function
The code is a bit complex, but it basically withdraws from a global tally to reduce the new item's price down to an initial floor (30% original), before applying an additional category-specific discount to reduce the price  even further (down to a minimum of 20% original).
YES I KNOW THE TESTS ARE FAILING. I had to comment out a line in `cs\production.go` that called the function because it was preventing me from running tests in the first place. I still dunno how to pass it a copy of the rules
Also fixed several bugs like incorrect miniaturization rounding & space dock not repairing enough